pyproject.toml
[tool.poetry]
name = "shelltools"
version = "0.0.0"
description = ""
authors = []
packages = [
    { include = "shelltools", from="src" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "2.1.0"
pynput = "1.7.6"
pyserial = "3.5"
terminology = "1.0.16"
unpadded = { git = "https://github.com/StarQTius/Unpadded", branch = "v1" }
numpy = "^1.26.0"

[tool.poetry.group.test.dependencies]
pytest = "7.4.2"
pytest-asyncio = "0.21.1"
pytest-timeout = "2.1.0"
mock_serial = "0.0.1"

[tool.poetry.group.lint.dependencies]
black = "23.3.0"
mypy = "1.3.0"
autoflake = "2.2.1"
isort = "5.12.0"

[tool.poetry.group.doc.dependencies]
sphinx = "7.2.6"
sphinx-markdown-builder = { git = "https://github.com/StarQTius/sphinx-markdown-builder" }

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core >= 1.5.2"]
build-backend = "poetry.core.masonry.api"
